Store ints as 63-bit tagged pointer
This is basically what Ruby does to make numbers more efficient. A
NatObject "pointer" ending with a 1 bit is actually a 63-bit integer
shifted to the left and and-ed with 1, like this: (n << 1) | 1

Here is a pretty good explation this technique:
https://stackoverflow.com/questions/33843393/how-does-ruby-differentiate-value-with-value-and-pointer/33843909#33843909

+++ src/hashmap.c
@@ -598,7 +598,8 @@ int hashmap_foreach(const struct hashmap *map,
 size_t hashmap_hash_string(const void *key)
 {
     const char *key_str = (const char *)key;
-    size_t hash = 0;
+    // FIXME: is this ok being a 32-bit number? more collisions?
+    int32_t hash = 0;
 
     for (; *key_str; ++key_str) {
         hash += *key_str;
